kill warning with cast.
git-svn-id: http://svn.research.cc.gatech.edu/kaos/ffs/trunk@16671 8ba6a8b9-845b-0410-94f0-974be2f5df47

+++ cod/tests/gray.c
@@ -91,7 +91,7 @@ void doOneTest(void (*f)(), int step)
 
 	/* first, the easy size check */
 	if (result != sizeof(gray)) { printf("result is %ld, not %ld\n", result,
-					     sizeof(gray));}
+					     (long)sizeof(gray));}
 	assert(result == sizeof(gray));
 
         /* then compare the whole image w/ expected gray */
